requirements.txt
asgiref==3.5.1
Django==3.2.13
django-debug-toolbar==3.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
